feat: list documents and assets with owner
