+++ src/satellite/src/db/mod.rs
@@ -1,3 +1,4 @@
 pub mod impls;
 pub mod store;
 pub mod types;
+pub mod utils;
