+++ src/satellite/src/storage/mod.rs
@@ -6,3 +6,4 @@ pub mod impls;
 pub mod store;
 pub mod types;
 pub mod url;
+pub mod utils;
